refactor: :art: Add getCurrentPath function to retrieve current page path
Add getCurrentPath function to retrieve current page path  This commit adds a new function called getCurrentPath, which retrieves the current path of the web page using the window.location.pathname property. This function will be useful in various parts of the application where we need to know the current path of the page.  By abstracting this functionality into a separate function, we can avoid duplicating code and make our code more modular and easier to maintain. Additionally, this function can be easily unit tested, which will help ensure that it works correctly in all cases.  Overall, this commit improves the codebase's organization and maintainability, and adds a useful utility function that will be used throughout the application.