refactor: :art: Add JSDoc comments to Route component
Add JSDoc comments to Route component  This commit adds JSDoc comments to the Route component, which was previously lacking documentation. The new comments describe the purpose of the function, its parameters, and its return value.  By adding JSDoc comments, we can better communicate the purpose and usage of the Route component to other developers who may work on this code in the future. This will help ensure that the component is used correctly and consistently throughout the application.  There are no changes to the functionality of the component itself, only the addition of comments.