refactor: :art: Add getCurrentPath utility function and JSDoc
Add getCurrentPath utility function and JSDoc comments to Router component  This commit adds a new utility function called getCurrentPath, which retrieves the current path of the web page using the window.location.pathname property. This function is used in the Router component to initialize the currentPath state value, and is also useful in other parts of the application where we need to know the current path.  In addition, this commit adds JSDoc comments to the Router component, which describe the purpose of the function, its parameters, its return value, and any potential errors that it may throw. These comments will be helpful in maintaining and extending the Router component in the future.  Overall, this commit improves the organization and documentation of the codebase, and adds a useful utility function that will be used throughout the application.